chore: fix preexisting mypy errors (windll guard, hint_seconds type)
resources.py used ctypes.windll unconditionally — attr-defined on non-Windows
mypy runs (runtime already fell back to sysconf via try/except). Guarded with
getattr. infer_periods_per_year took int | None but PerpConfig declares
bar_seconds_hint: float | None — widened the param. Both errors predated the
carry lane; CI lint's mypy step never ran green on main.

+++ src/quant_fund/backtest/perp_engine.py
@@ -83,7 +83,7 @@ class PerpBook:
         return sum(self.qty[s] * marks[s] for s in self.qty if s in marks)
 
 
-def infer_periods_per_year(times: list[datetime], hint_seconds: int | None = None) -> float:
+def infer_periods_per_year(times: list[datetime], hint_seconds: float | None = None) -> float:
     """Bar-spacing annualization on a 365.25-day calendar.
 
     Median spacing is robust to isolated missing bars; a hint wins only when
